Remade Controller and RestController, jsp pages, js files, validator, dao, ExceptionHandler for Order.class

+++ mvc-war/src/main/java/com/arttu/mvc/controller/OrderController.java
@@ -1,5 +1,10 @@
 package com.arttu.mvc.controller;
 
+import com.arttu.mvc.exception.client.ClientNotFoundException;
+import com.arttu.mvc.exception.item.ItemNotFoundException;
+import com.arttu.mvc.exception.order.OrderNotFoundException;
+import com.arttu.mvc.model.Client;
+import com.arttu.mvc.model.Item;
 import com.arttu.mvc.model.Order;
 import com.arttu.mvc.service.ClientService;
 import com.arttu.mvc.service.ItemService;
@@ -7,16 +12,14 @@ import com.arttu.mvc.service.OrderService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.servlet.ModelAndView;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 
-@RestController
+@Controller
 public class OrderController {
 
     private static final Logger LOGGER = LogManager.getLogger(OrderController.class);
@@ -34,79 +37,74 @@ public class OrderController {
     @GetMapping("/order")
     public ModelAndView department() {
         ModelAndView modelAndView = new ModelAndView("order/order");
-        modelAndView.addObject("orderList", orderService.findAll());
-        modelAndView.setStatus(HttpStatus.OK);
+        List<Order> orderList = orderService.findAll();
+        if (orderList == null) throw new OrderNotFoundException("Order not found");
+        modelAndView.addObject("orderList", orderList);
         return modelAndView;
     }
 
     @GetMapping("/order/add")
     public ModelAndView add() {
         ModelAndView modelAndView = new ModelAndView("order/addOrder");
-        modelAndView.addObject("itemList", itemService.findAll());
-        modelAndView.addObject("clientList", clientService.findAll());
-        modelAndView.setStatus(HttpStatus.OK);
+        List<Item> itemList = itemService.findAll();
+        List<Client> clientList = clientService.findAll();
+        if (itemList == null) throw new ItemNotFoundException("Item not found");
+        if (clientList == null) throw new ClientNotFoundException("Client not found");
+        modelAndView.addObject("itemList", itemList);
+        modelAndView.addObject("clientList", clientList);
         return modelAndView;
     }
 
     @GetMapping("/order/add/{id}")
     public ModelAndView add(@PathVariable int id) {
         ModelAndView modelAndView = new ModelAndView("order/addOrder");
-        modelAndView.addObject("item", itemService.findById(id));
-        modelAndView.addObject("itemList", itemService.findAll());
-        modelAndView.addObject("clientList", clientService.findAll());
-        modelAndView.setStatus(HttpStatus.OK);
+        Item item = itemService.findById(id);
+        List<Item> itemList = itemService.findAll();
+        List<Client> clientList = clientService.findAll();
+        if (item == null) throw new ItemNotFoundException("Item not found");
+        if (itemList == null) throw new ItemNotFoundException("Item not found");
+        if (clientList == null) throw new ClientNotFoundException("Client not found");
+        modelAndView.addObject("item", item);
+        modelAndView.addObject("itemList", itemList);
+        modelAndView.addObject("clientList", clientList);
         return modelAndView;
-    }
-
-    @PostMapping(value = "/order/add", produces = {MediaType.APPLICATION_JSON_VALUE},
-            consumes = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseEntity<Map<String, Object>> add(@RequestBody Order order) {
-        Map<String, Object> response = new HashMap<>();
-        HttpStatus status = HttpStatus.OK;
-        orderService.add(order);
-        response.put("stat", 1);
-        return new ResponseEntity<>(response, status);
     }
 
     @GetMapping(value = "/order/edit")
     public ModelAndView edit() {
         ModelAndView modelAndView = new ModelAndView("order/editOrder");
-        modelAndView.addObject("itemList", itemService.findAll());
-        modelAndView.addObject("clientList", clientService.findAll());
-        modelAndView.setStatus(HttpStatus.OK);
+        List<Item> itemList = itemService.findAll();
+        List<Client> clientList = clientService.findAll();
+        if (itemList == null) throw new ItemNotFoundException("Item not found");
+        if (clientList == null) throw new ClientNotFoundException("Client not found");
+        modelAndView.addObject("itemList", itemList);
+        modelAndView.addObject("clientList", clientList);
         return modelAndView;
     }
-
 
     @GetMapping(value = "/order/edit/{id}")
     public ModelAndView edit(@PathVariable int id) {
         ModelAndView modelAndView = new ModelAndView("order/editOrder");
-        modelAndView.addObject("clientList", clientService.findAll());
-        modelAndView.addObject("itemList", itemService.findAll());
-        modelAndView.addObject("orderList", orderService.findById(id));
-        modelAndView.setStatus(HttpStatus.OK);
+        List<Item> itemList = itemService.findAll();
+        List<Client> clientList = clientService.findAll();
+        Order order = orderService.findById(id);
+        if (itemList == null) throw new ItemNotFoundException("Item not found");
+        if (clientList == null) throw new ClientNotFoundException("Client not found");
+        if (order == null) throw new OrderNotFoundException("Order not found");
+        modelAndView.addObject("clientList", clientList);
+        modelAndView.addObject("itemList", itemList);
+        modelAndView.addObject("orderList", order);
         return modelAndView;
-    }
-
-    @PostMapping(value = "/order/edit", produces = {MediaType.APPLICATION_JSON_VALUE},
-            consumes = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseEntity<Map<String, Object>> edit(@RequestBody Order order) {
-        Map<String, Object> response = new HashMap<>();
-        HttpStatus status = HttpStatus.OK;
-        try {
-            orderService.edit(order);
-            response.put("stat", 1);
-        } catch (Exception e) {
-            LOGGER.error(e);
-            response.put("stat", 0);
-            status = HttpStatus.BAD_REQUEST;
-        }
-        return new ResponseEntity<>(response, status);
     }
 
     @GetMapping(value = "/order/delete/{id}")
     public ModelAndView delete(@PathVariable int id) {
-        orderService.deleteById(id);
+        Order order = orderService.findById(id);
+        if (order == null) {
+            throw new OrderNotFoundException("Order not found");
+        } else {
+            orderService.deleteById(id);
+        }
         return new ModelAndView("redirect:/order");
     }
 }
